# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: media
    description: Media processing jobs
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /media/jobs:
    get:
      operationId: listMediaJobs
      tags: [media]
      summary: List media processing jobs
      parameters:
        - name: archive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Media processing jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaJob"
    post:
      operationId: createMediaJob
      tags: [media]
      summary: Create a media processing job
      description: Uploads one media asset and starts a processing job.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MediaJobInput"
      responses:
        "202":
          description: Processing job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaJob"
        "400":
          description: Invalid upload or processing request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /media/jobs/{jobId}:
    get:
      operationId: getMediaJob
      tags: [media]
      summary: Get a media processing job
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Media processing job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaJob"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /media/jobs/{jobId}/events:
    get:
      operationId: streamMediaJobEvents
      tags: [media]
      summary: Stream job execution events
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Server-sent execution events
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /media/jobs/{jobId}/output:
    get:
      operationId: downloadMediaJobOutput
      tags: [media]
      summary: Download the completed media output
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Processed media file
          content:
            video/mp4:
              schema:
                type: string
                format: binary
            audio/mpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Output not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /media/jobs/{jobId}/subtitles:
    get:
      operationId: downloadMediaJobSubtitles
      tags: [media]
      summary: Download generated or attached captions
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Subtitle file
          content:
            application/x-subrip:
              schema:
                type: string
                format: binary
            text/vtt:
              schema:
                type: string
                format: binary
        "404":
          description: Captions not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /metrics/telemetry:
    get:
      operationId: getMediaTelemetry
      tags: [media]
      summary: Get media processing telemetry
      responses:
        "200":
          description: Current processing telemetry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaTelemetry"
  /integrations/status:
    get:
      operationId: getIntegrationDiagnostics
      tags: [media]
      summary: Verify external media integrations
      responses:
        "200":
          description: Sanitized provider connection diagnostics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationDiagnostics"
  /media/analytics:
    get:
      operationId: getMediaAnalytics
      tags: [media]
      summary: Get job analytics
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: codec
          in: query
          schema:
            type: string
        - name: preset
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Media job analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAnalytics"
components:
  parameters:
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    MediaJobInput:
      type: object
      properties:
        file:
          type: string
          format: binary
        preset:
          type: string
          enum:
            - vertical-reel
            - extract-audio
            - burn-subtitles
            - generate-subtitles
            - compress-video
            - upscale-video
            - custom
        prompt:
          type: string
        subtitle:
          type: string
          format: binary
          description: Optional SRT or VTT sidecar caption file. Required for burn-subtitles when subtitleMode is upload.
        subtitleMode:
          type: string
          enum: [upload, generate]
          description: For burn-subtitles, upload a caption file or generate captions from the source audio.
      required:
        - file
    MediaJob:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        status:
          type: string
          enum: [queued, processing, healing, succeeded, failed]
        preset:
          type: string
        prompt:
          type: string
        outputUrl:
          type: ["string", "null"]
        outputFilename:
          type: ["string", "null"]
        outputMimeType:
          type: ["string", "null"]
        subtitleSource:
          type: ["string", "null"]
          enum: [uploaded, generated, null]
        subtitleUrl:
          type: ["string", "null"]
        subtitleFilename:
          type: ["string", "null"]
        progressPercent:
          type: number
          minimum: 0
          maximum: 100
        stage:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
        attempt:
          type: number
        error:
          type: ["string", "null"]
        mediaInfo:
          $ref: "#/components/schemas/MediaInspection"
      required:
        - id
        - filename
        - status
        - preset
        - prompt
        - outputUrl
        - outputFilename
        - outputMimeType
        - subtitleSource
        - subtitleUrl
        - subtitleFilename
        - progressPercent
        - stage
        - createdAt
        - completedAt
        - attempt
        - error
        - mediaInfo
    MediaInspection:
      type: object
      properties:
        durationSeconds:
          type: number
        formatName:
          type: string
        sizeBytes:
          type: number
        hasVideo:
          type: boolean
        hasAudio:
          type: boolean
        videoCodec:
          type: ["string", "null"]
        audioCodec:
          type: ["string", "null"]
        streamCount:
          type: number
      required:
        - durationSeconds
        - formatName
        - sizeBytes
        - hasVideo
        - hasAudio
        - videoCodec
        - audioCodec
        - streamCount
    MediaTelemetry:
      type: object
      properties:
        activeJobs:
          type: number
        completedJobs:
          type: number
        failedJobs:
          type: number
        selfHealAttempts:
          type: number
        totalTokensUsed:
          type: number
        lastFfmpegError:
          type: string
        grafanaMcpConnected:
          type: boolean
        endpoint:
          type: string
      required:
        - activeJobs
        - completedJobs
        - failedJobs
        - selfHealAttempts
        - totalTokensUsed
        - grafanaMcpConnected
        - endpoint
    IntegrationProviderStatus:
      type: object
      properties:
        provider:
          type: string
          enum: [parallel, clickhouse, grafana]
        state:
          type: string
          enum: [connected, not_configured, error]
        lastCheckedAt:
          type: ["string", "null"]
          format: date-time
        lastSuccessAt:
          type: ["string", "null"]
          format: date-time
        lastError:
          type: string
        lastWriteAt:
          type: ["string", "null"]
          format: date-time
      required:
        - provider
        - state
        - lastCheckedAt
        - lastSuccessAt
    IntegrationDiagnostics:
      type: object
      properties:
        checkedAt:
          type: string
          format: date-time
        providers:
          type: array
          items:
            $ref: "#/components/schemas/IntegrationProviderStatus"
      required:
        - checkedAt
        - providers
    MediaAnalyticsRecord:
      type: object
      properties:
        jobId:
          type: string
        filename:
          type: string
        preset:
          type: string
        durationSeconds:
          type: number
        videoCodec:
          type: ["string", "null"]
        audioCodec:
          type: ["string", "null"]
        status:
          type: string
        attempts:
          type: number
        timestamp:
          type: string
      required:
        - jobId
        - filename
        - preset
        - durationSeconds
        - videoCodec
        - audioCodec
        - status
        - attempts
        - timestamp
    MediaAnalytics:
      type: object
      properties:
        totalRecords:
          type: number
        clickhouseMcpConnected:
          type: boolean
        records:
          type: array
          items:
            $ref: "#/components/schemas/MediaAnalyticsRecord"
      required:
        - totalRecords
        - clickhouseMcpConnected
        - records
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error